+++ xpserver/SDK/CHeaders/XPLM/XPLMDisplay.h
@@ -2,11 +2,11 @@
 #define _XPLMDisplay_h_
 
 /*
- * Copyright 2005 Sandy Barbour and Ben Supnik
+ * Copyright 2005-2012 Sandy Barbour and Ben Supnik
  * 
  * All rights reserved.  See license.txt for usage.
  * 
- * X-Plane SDK Version: 1.0.2                                                  
+ * X-Plane SDK Version: 2.1.1                                                  
  *
  */
 
@@ -716,7 +716,7 @@ XPLM_API void                 XPLMUnregisterHotKey(
  * Returns the number of current hot keys.                                     
  *
  */
-XPLM_API long                 XPLMCountHotKeys(void);
+XPLM_API int                  XPLMCountHotKeys(void);
 
 /*
  * XPLMGetNthHotKey
@@ -725,7 +725,7 @@ XPLM_API long                 XPLMCountHotKeys(void);
  *
  */
 XPLM_API XPLMHotKeyID         XPLMGetNthHotKey(
-                                   long                 inIndex);    
+                                   int                  inIndex);    
 
 /*
  * XPLMGetHotKeyInfo
