+++ xpserver/SDK/CHeaders/XPLM/XPLMNavigation.h
@@ -2,11 +2,11 @@
 #define _XPLMNavigation_h_
 
 /*
- * Copyright 2005 Sandy Barbour and Ben Supnik
+ * Copyright 2005-2012 Sandy Barbour and Ben Supnik
  * 
  * All rights reserved.  See license.txt for usage.
  * 
- * X-Plane SDK Version: 1.0.2                                                  
+ * X-Plane SDK Version: 2.1.1                                                  
  *
  */
 
@@ -206,7 +206,12 @@ XPLM_API XPLMNavRef           XPLMFindNavAid(
  * The buffer for IDs should be at least 6 chars and the buffer for names 
  * should be at least 41 chars, but since these values are likely to go up, I 
  * recommend passing at least 32 chars for IDs and 256 chars for names when 
- * possible.                                                                   
+ * possible. 
+ * 
+ * The outReg parameter tells if the navaid is within the local "region" of 
+ * loaded DSFs.  (This information may not be particularly useful to plugins.) 
+ * The parameter is a single byte value 1 for true or 0 for false, not a C 
+ * string.                                                                     
  *
  */
 XPLM_API void                 XPLMGetNavAidInfo(
@@ -243,7 +248,7 @@ XPLM_API void                 XPLMGetNavAidInfo(
  * This routine returns the number of entries in the FMS.                      
  *
  */
-XPLM_API long                 XPLMCountFMSEntries(void);
+XPLM_API int                  XPLMCountFMSEntries(void);
 
 /*
  * XPLMGetDisplayedFMSEntry
@@ -251,7 +256,7 @@ XPLM_API long                 XPLMCountFMSEntries(void);
  * This routine returns the index of the entry the pilot is viewing.           
  *
  */
-XPLM_API long                 XPLMGetDisplayedFMSEntry(void);
+XPLM_API int                  XPLMGetDisplayedFMSEntry(void);
 
 /*
  * XPLMGetDestinationFMSEntry
@@ -259,7 +264,7 @@ XPLM_API long                 XPLMGetDisplayedFMSEntry(void);
  * This routine returns the index of the entry the FMS is flying to.           
  *
  */
-XPLM_API long                 XPLMGetDestinationFMSEntry(void);
+XPLM_API int                  XPLMGetDestinationFMSEntry(void);
 
 /*
  * XPLMSetDisplayedFMSEntry
@@ -268,7 +273,7 @@ XPLM_API long                 XPLMGetDestinationFMSEntry(void);
  *  *
  */
 XPLM_API void                 XPLMSetDisplayedFMSEntry(
-                                   long                 inIndex);    
+                                   int                  inIndex);    
 
 /*
  * XPLMSetDestinationFMSEntry
@@ -277,7 +282,7 @@ XPLM_API void                 XPLMSetDisplayedFMSEntry(
  *
  */
 XPLM_API void                 XPLMSetDestinationFMSEntry(
-                                   long                 inIndex);    
+                                   int                  inIndex);    
 
 /*
  * XPLMGetFMSEntryInfo
@@ -292,11 +297,11 @@ XPLM_API void                 XPLMSetDestinationFMSEntry(
  *
  */
 XPLM_API void                 XPLMGetFMSEntryInfo(
-                                   long                 inIndex,    
+                                   int                  inIndex,    
                                    XPLMNavType *        outType,    /* Can be NULL */
                                    char *               outID,    /* Can be NULL */
                                    XPLMNavRef *         outRef,    /* Can be NULL */
-                                   long *               outAltitude,    /* Can be NULL */
+                                   int *                outAltitude,    /* Can be NULL */
                                    float *              outLat,    /* Can be NULL */
                                    float *              outLon);    /* Can be NULL */
 
@@ -310,9 +315,9 @@ XPLM_API void                 XPLMGetFMSEntryInfo(
  *
  */
 XPLM_API void                 XPLMSetFMSEntryInfo(
-                                   long                 inIndex,    
+                                   int                  inIndex,    
                                    XPLMNavRef           inRef,    
-                                   long                 inAltitude);    
+                                   int                  inAltitude);    
 
 /*
  * XPLMSetFMSEntryLatLon
@@ -322,10 +327,10 @@ XPLM_API void                 XPLMSetFMSEntryInfo(
  *
  */
 XPLM_API void                 XPLMSetFMSEntryLatLon(
-                                   long                 inIndex,    
+                                   int                  inIndex,    
                                    float                inLat,    
                                    float                inLon,    
-                                   long                 inAltitude);    
+                                   int                  inAltitude);    
 
 /*
  * XPLMClearFMSEntry
@@ -335,7 +340,7 @@ XPLM_API void                 XPLMSetFMSEntryLatLon(
  *
  */
 XPLM_API void                 XPLMClearFMSEntry(
-                                   long                 inIndex);    
+                                   int                  inIndex);    
 
 /***************************************************************************
  * GPS RECEIVER
