+++ xpserver/SDK/CHeaders/XPLM/XPLMPlanes.h
@@ -2,11 +2,11 @@
 #define _XPLMPlanes_h_
 
 /*
- * Copyright 2005 Sandy Barbour and Ben Supnik
+ * Copyright 2005-2012 Sandy Barbour and Ben Supnik
  * 
  * All rights reserved.  See license.txt for usage.
  * 
- * X-Plane SDK Version: 1.0.2                                                  
+ * X-Plane SDK Version: 2.1.1                                                  
  *
  */
 
@@ -80,7 +80,7 @@ XPLM_API void                 XPLMPlaceUserAtAirport(
  */
 typedef struct {
      /* The size of the draw state struct.                                          */
-     long                      structSize;
+     int                       structSize;
      /* A ratio from [0..1] describing how far the landing gear is extended.        */
      float                     gearPosition;
      /* Ratio of flap deployment, 0 = up, 1 = full deploy.                          */
